Add files via upload
This is just a tiny test case made from a portion of chromosome 1 and M. fermentans. I'm not able to even upload the limited data (the zipped file is 1.8 GB). I get the message that the upload limit is 25 MB. We should discuss a better way to share/store data.